+++ cime_config/cesm/config_files.xml
@@ -43,7 +43,8 @@
 
   <entry id="COMPILERS_SPEC_FILE">
     <type>char</type>
-    <default_value>$CIMEROOT/cime_config/$MODEL/machines/config_compilers.xml</default_value>
+    <default_value>$CIMEROOT/cime_config/$MODEL/machines/config_build.xml</default_value>
+    <schema version="2.0">$CIMEROOT/cime_config/xml_schemas/config_build.xsd</schema>
     <group>case_last</group>
     <file>env_case.xml</file>
     <desc>file containing compiler specifications for target model primary component (for documentation only - DO NOT EDIT)</desc>
@@ -332,4 +333,3 @@
   </entry>
 
 </entry_id>
-
