+++ cime_config/cesm/machines/config_build.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<config_build>
+<config_compilers version="2.0">
 <!--
 ========================================================================
 This file defines compiler flags for building CESM.  General flags are listed first
@@ -1367,4 +1367,4 @@ for mct, etc.
   </SLIBS>
 </compiler>
 
-</config_build>
+</config_compilers>
